add retries

+++ dags/jobs/file_downloader.py
@@ -3,7 +3,7 @@ import concurrent.futures
 import requests
 import os
 
-
+MAX_RETRIES = 1
 URL = 'http://snap.stanford.edu/data/amazon/productGraph/categoryFiles/meta_Movies_and_TV.json.gz'
 # OUTPUT = 'amazon/video.json.gz'
 OUTPUT = '/usr/local/spark/resources/data/meta_movies_and_tv/meta_movies_and_tv.json.gz'
@@ -15,14 +15,26 @@ async def get_size(url):
     return size
 
 
-def download_range(url, start, end, output):
-    headers = {'Range': f'bytes={start}-{end}'}
-    response = requests.get(url, headers=headers)
+def download_range(url, start, end, output, retries = 0):
+    try:
+        if end > 10 * 1000000 and end < 90 * 1000000:
+            test = 1/0
+        headers = {'Range': f'bytes={start}-{end}'}
+        response = requests.get(url, headers=headers)
 
-    with open(output, 'wb') as f:
-        for part in response.iter_content(1024):
-            f.write(part)
+        with open(output, 'wb') as f:
+            for part in response.iter_content(1024):
+                f.write(part)
+        return True
+    except Exception as e:
+        print(e, retries, end, '\n')
+        if retries < MAX_RETRIES:
+            download_range(url, start, end, output, retries+1)
+        else:
+            raise e
 
+def is_success(future):
+   return future.done() and not future.cancelled() and future.exception() is None
 
 async def download(executor, url, output, chunk_size=1000000):
     loop = asyncio.get_event_loop()
@@ -42,16 +54,22 @@ async def download(executor, url, output, chunk_size=1000000):
         for i, start in enumerate(chunks)
     ]
 
-    await asyncio.wait(tasks)
+    done, pending = await asyncio.wait(tasks, return_when=asyncio.FIRST_EXCEPTION)
 
-    with open(output, 'wb') as o:
-        for i in range(len(chunks)):
-            chunk_path = f'{output}.part{i}'
+    if all(is_success(f) for f in done):
+        with open(output, 'wb') as o:
+            for i in range(len(chunks)):
+                chunk_path = f'{output}.part{i}'
 
-            with open(chunk_path, 'rb') as s:
-                o.write(s.read())
+                with open(chunk_path, 'rb') as s:
+                    o.write(s.read())
 
-            os.remove(chunk_path)
+                os.remove(chunk_path)
+    else:
+        for f in pending:
+            f.cancel()
+        failed_future = next(f for f in done if not is_success(f))
+        raise failed_future.exception()
 
 def parallel_download():
     executor = concurrent.futures.ThreadPoolExecutor(max_workers=4)
